requirements.txt
requests 
beautifulsoup4
selenium
python-dotenv
python-telegram-bot